🌟 enhancement(VSecM Sentinel): code cleanup and manifest generation
