+++ app/sentinel/busywait/main.go
@@ -11,117 +11,11 @@
 package main
 
 import (
-	"bufio"
-	"context"
-	"github.com/vmware-tanzu/secrets-manager/app/sentinel/internal/safe"
-	entity "github.com/vmware-tanzu/secrets-manager/core/entity/data/v1"
-	"github.com/vmware-tanzu/secrets-manager/core/env"
+	"github.com/vmware-tanzu/secrets-manager/app/sentinel/busywait/init"
 	"github.com/vmware-tanzu/secrets-manager/core/log"
 	"github.com/vmware-tanzu/secrets-manager/core/probe"
 	"github.com/vmware-tanzu/secrets-manager/core/system"
-	"os"
-	"strconv"
-	"strings"
-	"time"
 )
-
-const delimiter = "--"
-const separator = ":"
-
-type command string
-
-const workload command = "w"
-const namespace command = "n"
-const secret command = "s"
-const transformation command = "t"
-const sleep = "sleep"
-
-func processCommandBlock(ctx context.Context, sc entity.SentinelCommand) {
-	// TODO: change the signature of the function to accept the context.
-	safe.Post(ctx, sc)
-}
-
-func doSleep(seconds int) {
-	time.Sleep(time.Duration(seconds) * time.Millisecond)
-}
-
-func executeInitCommand() {
-	cid := "VSECMSENTINEL"
-	filePath := env.SentinelInitCommandPath()
-	file, err := os.Open(filePath)
-	if err != nil {
-		log.InfoLn(
-			&cid,
-			"no initialization file found… skipping custom initialization.",
-		)
-		return
-	}
-	defer func(file *os.File) {
-		err := file.Close()
-		if err != nil {
-			log.ErrorLn(&cid, "Error closing initialization file: ", err.Error())
-		}
-	}(file)
-
-	ctx := context.Background()
-
-	scanner := bufio.NewScanner(file)
-	var sc entity.SentinelCommand
-
-	for scanner.Scan() {
-		line := strings.TrimSpace(scanner.Text())
-
-		if line == "" {
-			continue
-		}
-
-		parts := strings.SplitN(line, separator, 2)
-
-		if len(parts) != 2 && line != delimiter {
-			continue
-		}
-
-		if line == delimiter {
-			if sc.ShouldSleep {
-				doSleep(sc.SleepIntervalMs)
-				sc = entity.SentinelCommand{}
-				continue
-			}
-
-			processCommandBlock(ctx, sc)
-			sc = entity.SentinelCommand{}
-			continue
-		}
-
-		key := parts[0]
-		value := parts[1]
-
-		switch command(key) {
-		case workload:
-			sc.WorkloadId = value
-		case namespace:
-			sc.Namespace = value
-		case secret:
-			sc.Secret = value
-		case transformation:
-			sc.Template = value
-		case sleep:
-			sc.ShouldSleep = true
-			intms, _ := strconv.Atoi(value)
-			sc.SleepIntervalMs = intms
-		default:
-			log.InfoLn(&cid, "unknown command: ", key)
-		}
-	}
-
-	if err := scanner.Err(); err != nil {
-		log.ErrorLn(
-			&cid,
-			"Error reading initialization file: ",
-			err.Error(),
-		)
-	}
-}
 
 func main() {
 	id := "VSECMSENTINEL"
@@ -134,7 +28,12 @@ func main() {
 	log.PrintEnvironmentInfo(&id, envVarsToPrint)
 
 	log.InfoLn(&id, "Executing the initialization commands (if any)")
-	executeInitCommand()
+	// Execute the initialization commands (if any)
+	// This overloads the functionality of this process.
+	// If we end up adding more functionality to this process,
+	// we should refactor this and create a new process for the
+	// new functionality.
+	init.RunInitCommands()
 	log.InfoLn(&id, "Initialization commands executed successfully")
 
 	// Run on the main thread to wait forever.
