+++ helm-charts/0.22.3/charts/safe/values.yaml
@@ -21,7 +21,7 @@ environments:
   - name: SPIFFE_ENDPOINT_SOCKET
     value: "unix:///spire-agent-socket/agent.sock"
   - name: VSECM_LOG_LEVEL
-    value: "7"
+    value: "3"
   - name: VSECM_PROBE_LIVENESS_PORT
     value: ":8081"
   - name: VSECM_PROBE_READINESS_PORT
